+++ elements/storytelling/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.1.0](https://github.com/EOX-A/EOxElements/compare/storytelling-v0.0.0...storytelling-v0.1.0) (2024-02-27)
+
+
+### Features
+
+* Add markdown rendering via `markdown-it` as string and URL ([#690](https://github.com/EOX-A/EOxElements/issues/690)) ([42154a3](https://github.com/EOX-A/EOxElements/commit/42154a3c133ee29bfdfedc8816730ea859039419))
+* Render markdown from default slot ([#696](https://github.com/EOX-A/EOxElements/issues/696)) ([cfef85a](https://github.com/EOX-A/EOxElements/commit/cfef85a6675a58e9637e684c2edd48e8c42f9f39))
+
+
+### Bug Fixes
+
+* **docs:** MarkdownURL pattern ([#697](https://github.com/EOX-A/EOxElements/issues/697)) ([edbc1ca](https://github.com/EOX-A/EOxElements/commit/edbc1ca21c5bd3d8fe0a2b4d918566a72fbc40d3))
+
 ## 0.0.0 (2024-02-22)